抽离逻辑代码
