+++ src/hooks/useFileteredTodos.js
@@ -1,0 +1,19 @@
+import { ref, computed } from "vue";
+
+export default function useFileteredTodos(todos) {
+  const filter = ref("all");
+  const filteredTodos = computed(() => {
+    switch (filter.value) {
+      case "done":
+        return todos.value.filter((todo) => todo.completed);
+      case "todo":
+        return todos.value.filter((todo) => !todo.completed);
+      default:
+        return todos.value;
+    }
+  });
+  return {
+    filter,
+    filteredTodos,
+  };
+}
